fonctionnalité de supression des chansons
