+++ package.json
@@ -1,62 +1,62 @@
 {
-  "name": "yatga",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~11.2.14",
-    "@angular/cdk": "^11.2.13",
-    "@angular/common": "~11.2.14",
-    "@angular/compiler": "~11.2.14",
-    "@angular/core": "~11.2.14",
-    "@angular/fire": "^6.1.5",
-    "@angular/forms": "~11.2.14",
-    "@angular/material": "^11.2.13",
-    "@angular/platform-browser": "~11.2.14",
-    "@angular/platform-browser-dynamic": "~11.2.14",
-    "@angular/router": "~11.2.14",
-    "@fortawesome/fontawesome-free": "^5.15.3",
-    "@tableaubits/hang": "^1.4.4",
-    "@types/lodash": "^4.14.171",
-    "firebase": "^7.0 || ^8.0",
-    "lodash": "^4.17.21",
-    "mdb-angular-ui-kit": "^1.0.0-beta8",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.3"
-  },
-  "devDependencies": {
-    "@angular-devkit/architect": ">= 0.900 < 0.1300",
-    "@angular-devkit/build-angular": "~0.1102.13",
-    "@angular/cli": "~11.2.14",
-    "@angular/compiler-cli": "~11.2.14",
-    "@types/events": "^3.0.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "firebase-tools": "^8.0.0 || ^9.0.0",
-    "fuzzy": "^0.1.3",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.0.1",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "jsonc-parser": "^3.0.0",
-    "karma": "~6.1.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.5.0",
-    "open": "^7.0.3",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
-  }
+	"name": "yatga",
+	"version": "0.0.0",
+	"scripts": {
+		"ng": "ng",
+		"start": "ng serve",
+		"build": "ng build",
+		"test": "ng test",
+		"lint": "ng lint",
+		"e2e": "ng e2e"
+	},
+	"private": true,
+	"dependencies": {
+		"@angular/animations": "~11.2.14",
+		"@angular/cdk": "^11.2.13",
+		"@angular/common": "~11.2.14",
+		"@angular/compiler": "~11.2.14",
+		"@angular/core": "~11.2.14",
+		"@angular/fire": "^6.1.5",
+		"@angular/forms": "~11.2.14",
+		"@angular/material": "^11.2.13",
+		"@angular/platform-browser": "~11.2.14",
+		"@angular/platform-browser-dynamic": "~11.2.14",
+		"@angular/router": "~11.2.14",
+		"@fortawesome/fontawesome-free": "^5.15.3",
+		"@tableaubits/hang": "1.4.4",
+		"@types/lodash": "^4.14.171",
+		"firebase": "^7.0 || ^8.0",
+		"lodash": "^4.17.21",
+		"mdb-angular-ui-kit": "^1.0.0-beta8",
+		"rxjs": "~6.6.0",
+		"tslib": "^2.0.0",
+		"zone.js": "~0.11.3"
+	},
+	"devDependencies": {
+		"@angular-devkit/architect": ">= 0.900 < 0.1300",
+		"@angular-devkit/build-angular": "~0.1102.13",
+		"@angular/cli": "~11.2.14",
+		"@angular/compiler-cli": "~11.2.14",
+		"@types/events": "^3.0.0",
+		"@types/jasmine": "~3.6.0",
+		"@types/node": "^12.11.1",
+		"codelyzer": "^6.0.0",
+		"firebase-tools": "^8.0.0 || ^9.0.0",
+		"fuzzy": "^0.1.3",
+		"inquirer": "^6.2.2",
+		"inquirer-autocomplete-prompt": "^1.0.1",
+		"jasmine-core": "~3.6.0",
+		"jasmine-spec-reporter": "~5.0.0",
+		"jsonc-parser": "^3.0.0",
+		"karma": "~6.1.0",
+		"karma-chrome-launcher": "~3.1.0",
+		"karma-coverage": "~2.0.3",
+		"karma-jasmine": "~4.0.0",
+		"karma-jasmine-html-reporter": "~1.5.0",
+		"open": "^7.0.3",
+		"protractor": "~7.0.0",
+		"ts-node": "~8.3.0",
+		"tslint": "~6.1.0",
+		"typescript": "~4.1.5"
+	}
 }
