+++ src/app/components/constitution-page/constitution.component.ts
@@ -5,135 +5,135 @@ import { ActivatedRoute } from '@angular/router';
 import { Constitution, createMessage, CstReqGet, CstResUpdate, CstSongReqGetAll, CstSongResUpdate, EventType, extractMessageData, Message, Role, Song, User, UsrReqGet, UsrReqUnsubscribe, UsrResUpdate } from '@tableaubits/hang';
 import { AuthService } from 'src/app/services/auth.service';
 import { EMPTY_CONSTITUTION, OWNER_INDEX } from 'src/app/types/constitution';
-import { ManageSongsComponent } from '../manage-songs/manage-songs.component';
+import { ManageSongsComponent } from './manage-songs/manage-songs.component';
 
 enum ConstitutionSection {
-  SONG_LIST,
-  VOTES,
-  OWNER,
-  RESULTS,
-  EXPORT
+	SONG_LIST,
+	VOTES,
+	OWNER,
+	RESULTS,
+	EXPORT
 }
 
 @Component({
-  selector: 'app-constitution',
-  templateUrl: './constitution.component.html',
-  styleUrls: ['./constitution.component.scss']
+	selector: 'app-constitution',
+	templateUrl: './constitution.component.html',
+	styleUrls: ['./constitution.component.scss']
 })
 export class ConstitutionComponent {
 
-  private cstID;
-  constitution: Constitution;
-  users: Map<string, User>;
-  currentSection: ConstitutionSection;
-  songs: Map<number, Song>;
+	private cstID;
+	constitution: Constitution;
+	users: Map<string, User>;
+	currentSection: ConstitutionSection;
+	songs: Map<number, Song>;
 
-  constructor(
-    private auth: AuthService, 
-    private route: ActivatedRoute,
-    private dialog: MatDialog
-  ){
-    this.cstID = "";
-    this.constitution = EMPTY_CONSTITUTION;
-    this.currentSection = ConstitutionSection.SONG_LIST;
-    this.users = new Map();
-    this.songs = new Map();
-    this.auth.waitForAuth(this.handleEvents, this.onConnect, this);
-  }
+	constructor(
+		private auth: AuthService,
+		private route: ActivatedRoute,
+		private dialog: MatDialog
+	) {
+		this.cstID = "";
+		this.constitution = EMPTY_CONSTITUTION;
+		this.currentSection = ConstitutionSection.SONG_LIST;
+		this.users = new Map();
+		this.songs = new Map();
+		this.auth.waitForAuth(this.handleEvents, this.onConnect, this);
+	}
 
-  // HTML can't access the ConstiutionSection enum directly
-  public get constitutionSection(): typeof ConstitutionSection {
-    return ConstitutionSection; 
-  }
+	// HTML can't access the ConstiutionSection enum directly
+	public get constitutionSection(): typeof ConstitutionSection {
+		return ConstitutionSection;
+	}
 
-  private onConnect(): void {
-    this.route.params.subscribe((params) => {
-      this.cstID = params.cstID;  // TODO : check if user is a member of the constitution and redirect him
+	private onConnect(): void {
+		this.route.params.subscribe((params) => {
+			this.cstID = params.cstID;  // TODO : check if user is a member of the constitution and redirect him
 
-      const getCSTMessage = createMessage<CstReqGet>(EventType.CST_get, {ids: [this.cstID]})
-      this.auth.ws.send(getCSTMessage);
-    })
-  }
+			const getCSTMessage = createMessage<CstReqGet>(EventType.CST_get, { ids: [this.cstID] })
+			this.auth.ws.send(getCSTMessage);
+		})
+	}
 
-  private handleEvents(event: MessageEvent<any>): void {
-    let message = JSON.parse(event.data.toString()) as Message<unknown>;
-    
-    switch (message.event) {
-      case EventType.CST_update: {
-        const data = extractMessageData<CstResUpdate>(message).cstInfo;
-        if (data.id === this.cstID) {
-          this.constitution = data;
+	private handleEvents(event: MessageEvent<any>): void {
+		let message = JSON.parse(event.data.toString()) as Message<unknown>;
 
-          const newUsers = this.constitution.users.filter((uid) => !this.users.has(uid));
-          const unusedListens = Array.from(this.users.values()).filter((user) => !this.constitution.users.includes(user.uid)).map((user) => user.uid);
-          for (const uid of unusedListens) {
-            this.users.delete(uid)
-          }
+		switch (message.event) {
+			case EventType.CST_update: {
+				const data = extractMessageData<CstResUpdate>(message).cstInfo;
+				if (data.id === this.cstID) {
+					this.constitution = data;
 
-          const getUsersMessage = createMessage<UsrReqGet>(EventType.USER_get, {uids: newUsers})
-          this.auth.ws.send(getUsersMessage);
-          const unsubscribeUsersMessage = createMessage<UsrReqUnsubscribe>(EventType.USER_unsubscribe, {uids: unusedListens});
-          this.auth.ws.send(unsubscribeUsersMessage);
-          const getAllSongsMessage = createMessage<CstSongReqGetAll>(EventType.CST_SONG_get_all, {cstId: this.cstID, uid: ''}); // TODO : Remove uid for get all in Hang
-          this.auth.ws.send(getAllSongsMessage);
-        } 
-      }
-        break;
-      case EventType.USER_update: {
-        const data = extractMessageData<UsrResUpdate>(message).userInfo;
-        this.users.set(data.uid, data)
-      }
-        break;
-      case EventType.CST_SONG_update: {
-        // TODO : check if is correct constitution ?
-        const data = extractMessageData<CstSongResUpdate>(message);
-        this.songUpdate(data);
-      }
-        break;
-    }
-  }
+					const newUsers = this.constitution.users.filter((uid) => !this.users.has(uid));
+					const unusedListens = Array.from(this.users.values()).filter((user) => !this.constitution.users.includes(user.uid)).map((user) => user.uid);
+					for (const uid of unusedListens) {
+						this.users.delete(uid)
+					}
 
-  private songUpdate(response: CstSongResUpdate) {
-    const songInfo = response.songInfo;
-    switch (response.status) {
-      case "added" || "modified":
-        this.songs.set(songInfo.id, songInfo);
-        break;
-      case "removed":
-        this.songs.delete(songInfo.id);
-        break;
-    }
-  }
+					const getUsersMessage = createMessage<UsrReqGet>(EventType.USER_get, { uids: newUsers })
+					this.auth.ws.send(getUsersMessage);
+					const unsubscribeUsersMessage = createMessage<UsrReqUnsubscribe>(EventType.USER_unsubscribe, { uids: unusedListens });
+					this.auth.ws.send(unsubscribeUsersMessage);
+					const getAllSongsMessage = createMessage<CstSongReqGetAll>(EventType.CST_SONG_get_all, { cstId: this.cstID, uid: '' }); // TODO : Remove uid for get all in Hang
+					this.auth.ws.send(getAllSongsMessage);
+				}
+			}
+				break;
+			case EventType.USER_update: {
+				const data = extractMessageData<UsrResUpdate>(message).userInfo;
+				this.users.set(data.uid, data)
+			}
+				break;
+			case EventType.CST_SONG_update: {
+				// TODO : check if is correct constitution ?
+				const data = extractMessageData<CstSongResUpdate>(message);
+				this.songUpdate(data);
+			}
+				break;
+		}
+	}
 
-  openDialogManageSongs(): void {
-    const config = new MatDialogConfig();
+	private songUpdate(response: CstSongResUpdate) {
+		const songInfo = response.songInfo;
+		switch (response.status) {
+			case "added" || "modified":
+				this.songs.set(songInfo.id, songInfo);
+				break;
+			case "removed":
+				this.songs.delete(songInfo.id);
+				break;
+		}
+	}
 
-    config.data = {
-      cstID: this.cstID,
-      songs: this.songs
-    }
+	openDialogManageSongs(): void {
+		const config = new MatDialogConfig();
 
-    this.dialog.open(ManageSongsComponent, config);
-  }
+		config.data = {
+			cstID: this.cstID,
+			songs: this.songs
+		}
 
-  setCurrentSection(newSection: ConstitutionSection): void {
-    this.currentSection = newSection;
-  }
+		this.dialog.open(ManageSongsComponent, config);
+	}
 
-  isSectionActive(section: ConstitutionSection): boolean {
-    return section === this.currentSection;
-  } 
+	setCurrentSection(newSection: ConstitutionSection): void {
+		this.currentSection = newSection;
+	}
 
-  isOwner(user?: User): boolean {
-    if (user) return user.uid === this.constitution.users[OWNER_INDEX];
-    else return this.auth.uid === this.constitution.users[OWNER_INDEX];
-  }
-  
-  isAdmin(user: User): boolean {
-    return user.roles.includes(Role.ADMIN);
-  }
+	isSectionActive(section: ConstitutionSection): boolean {
+		return section === this.currentSection;
+	}
 
-  getUsers(): User[] {
-    return Array.from(this.users.values());
-  }
+	isOwner(user?: User): boolean {
+		if (user) return user.uid === this.constitution.users[OWNER_INDEX];
+		else return this.auth.uid === this.constitution.users[OWNER_INDEX];
+	}
+
+	isAdmin(user: User): boolean {
+		return user.roles.includes(Role.ADMIN);
+	}
+
+	getUsers(): User[] {
+		return Array.from(this.users.values());
+	}
 }
